Split commands file into routing and commands to make it more maintainable
